requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
PyYAML==6.0.1
celery==5.3.0
redis==4.5.0
python-dotenv==1.0.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.5

drf-yasg==1.21.7